A reimplementation of the Slate components for the Translator.
